+++ GUI/DataManager/DataManagerFrame.lua
@@ -4,8 +4,9 @@ addon.GUI.DataManagerFrame = {};
 local dataManagerFrame = addon.GUI.DataManagerFrame;
 
 function dataManagerFrame:Load()
+	local frame;
 	if addon.IsNotWotLKClassic() then
-		local frame = CreateFrame("Frame", "KrowiAF_DataManagerFrame", UIParent, "KrowiAF_DataManagerFrame_Template");
+		frame = CreateFrame("Frame", "KrowiAF_DataManagerFrame", UIParent, "KrowiAF_DataManagerFrame_Template");
 	end
 
 	addon.GUI.DataManagerFrame = frame;