+++ CHANGELOG.md
@@ -5,10 +5,12 @@ All notable changes to this project will be documented in this file.
 ### Added
 - Data manager that can be opened by clicking on the achievement points in the header to show/hide characters in/from the header tooltip
 - Dynamic obtainability (replaces the previous yes/no obtainable flag which opens new possibilities) - NOT ALL DATA UPDATED YET
+- Fated Raids Calendar events and reminders
 
 ### WotLK Classic
 - First port from retail
 - All features except event reminders, world map button and data manager are ported
+- Classic and TBC achievements and categories adjusted and dynamic obtainability added
 
 ## 9.2.5.37.0 - 2022-08-01
 ### Added